requirements.txt
easyocr
elasticsearch
opencv-python
openpyxl
PyMuPDF
pytesseract
python-docx
torch
docx2txt